Decision tree making and Prediction model creating
In this section, first we load the iris data set, then Divide the sample into a training data set with 2/3 of the observations.  secondly, we make a  decision tree for the training data set. At last we make a prediction model for test data set and look for predicting values and constructed the confusion matrix.