Multiple Linear Regression with dummy variable.
(a) Create a scatter plot with shelf on the vertical axis and rating on
the horizontal axis. Interpret the scatter plot.
(b) Create two dummy variables called shelf1 and shelf2 as follows:

(c) Now use rating as response variable and sugar, fiber, shelf1 and shelf2
as independent variables in a multiple linear regression model. What
is the estimated regression model for cereals located on the first shelf?
What is the estimated regression model for cereals located on the
second shelf? What is the estimated regression model for cereals
located on the third shelf?