Simple Logistic Regression for Churn data set Voicemail Plan
2. The data set churn contains customer information for a telecommunication
company. We are interested in predicting whether a customer would leave
the company’s service (=churn) or not based on a set of predictor variables.
In this task, we are going to perform a simple logistic regression using just
one predictor variable: VoiceMail Plan. This variable is a flag variable
indicating membership in a special plan.
(a) Read the data set churn into R, for example with the command
read.csv.

(b) Perform a simple logistic regression estimation with Churn. as response variable and VoiceMail Plan as predictor. What is the estimated logit for your model?


(c) Determine if the variable VMail.Plan is significant for predicting the
probability of churn for a customer.
...
(d) What is the probability of churning for a customer not belonging to
the VMail.Plan?