Multiple Linear Regression with Churn data set
3. Use the churn data set from the previous exercise to perform a multiple
logistic regression. The aim is still to estimate the probability that a
customer will churn based on some predictor variables.
(a) Perform a multiple logistic regression analysis with Churn. as response variable and Day Minutes, Evening Minutes, Night Minutes,
International Minutes, International Plan, VMail.Plan and CustServ.Calls as predictors. CustServ.Calls is a variable indicating whether
or not a customer had a high level of service calls (=no of service
calls). International Plan and VMail.Plan are flag variables telling
if the customer belongs to these plans or not. (Note that the names
of the variables are slightly different in the data set.)

(b) What is the estimated logit for this model?

(c) What is the probability of churning for a customer with 100 minutes for
each of day, evening and night minutes, no international minutes, belonging
to no plans and with no customer service calls?