Decision Trees making on iris data set
1. We are going to build a decision tree model for prediction of the variable
Species for the dataset iris.
(a) Make sure that the packages rpart and rpart.plot are loaded and that
the dataset iris is loaded.
(b) Separate the data set into a training data set with 2/3 of the observations and a validation data set with 1/3 of the observations.
(c) Create a standard decision tree based on the training data set and
create predictions for the validation data set.
(d) Give a summary of the training and the validation data.
(e) Construct the confusion matrix. Interpretation?
(f) What are the current values for the hyperparameters?
(g) Draw the decision tree.