Decision Tree  training and validation and remove variable
2. The data set churn contains customer information for a telecommunication
company. We are interested in predicting whether a customer would leave
the company’s service (=churn) or not based on a set of predictor variables.
In this task, we are going to use a decision tree model to predict churn.
(a) Read the data set churn into R, for example with the commands
read.csv and make sure that the packages rpart and rpart.plot are
loaded.
(b) Partition the data into two datasets: One should be used for training
and one should be used for validation of the model. The training data
set should contain 70% of the original observations.
(c) Create a decision tree based on the training data. Look at the tree.
What can you discover?
(d) Remove the variable Phone.