Neural Network weight and Unit conversion using nnet
1. The data set churn contains customer information for a telecommunication company. We are interested in predicting whether a customer would
leave the company’s service (=churn) or not based on a set of predictor
variables. In this task, we are going to construct a neural network model
for predicting churn (but to make the exercise a bit shorter, we just construct the model and don’t perform predictions, which would be the next
step).
(a) Read the data set churn into R, for example with the commands
read.csv and make sure that the package nnet is loaded.
(b) Partition the data into two datasets: One should be used for training
and one should be used for validation of the model. The training data
set should contain 70% of the original observations.
(c) Use the training data to construct a neural network with the function
nnet in the package nnet. You should use Churn. as target variable
and all other variables in the training dataset as input variables.
Use 2 units in the hidden layer.
(d) Which problem did you encounter in c)?
(e) Use the training data to construct a neural network with the function
nnet. You should use Churn. as target variable and the variables
Day.Mins and VMail.Plan as input variables. Use a maximum of
200 iterations and use 2 units in the hidden layer.
(f) Did the algorithm converge in e)? If yes, how many iterations were
required?
How many weights are used in your neural network? Is this the
number of weights that you also expect from theory (with one layer
and two hidden units)?
(g) Use the function summary to see the final values of the weights. Draw
(hand-written if you want) the structure of the neural network and
include the estimated final weights at your network connections.
(h) Increase the number of hidden units to 20 and train your network
once more. Did the algorithm converge? How many units do you
have now?