Cluster analysis with Kmeans function
1. Use the k-means algorithm to find three clusters for the following observations (just one variable!) per hand:
1 1.1 5 5.1 1.5 5.2 7.9 1.2 8.1 9
2. Use the same data set with 10 observations as in the theoretical exercise above.
Now, use the function kmeans in the package amap to perform a k-means clustering.
(a) Use the function plot to plot the data (without clusters).
(b) Use the function kmeans to cluster the data.
(c) As a result of the cluster analysis with the function kmeans you get (among
others) the components cluster, centers, withinss and size. Explain the
theoretical meaning (with your own words) of these components.
(d) Use the function plot to plot the clusters (with different colors for different
clusters).
(e) Use the function points to plot the cluster centers as well in the plot in d).
(f) The starting clusters in the k-means algorithm are randomly chosen. However, the final clustering often depends on the random seed which is used to
randomly choose the starting clusters (unfortunately). The random seed
can in R be set with the command set.seed(yournumber). Try some different values of the random seed and see if you always get the same clustering
or not.