Association Rules for Subset
(g) Use inspect to interpret the 20th rule generated (from the top) in the
created rules dataset. What does the lift value of this rule mean?
Interpret!
(h) Sort the generated rules w.r.t. lift (highest lift at the top). List the
ten rules with the highest lift values.
(i) Subset the generated rules in such a way that you only keep rules
with at least ”pork” in the left-hand-side of the rule and a lift value
higher than 1.2. How many rules are now kept? Which are the rules?